+++ packages/fhir_r4_db/CHANGELOG.md
@@ -1,5 +1,9 @@
 # fhir_r4_db
 
+## [0.4.0]
+
+* just update to use the most recent fhir_r4 package
+
 ## [0.3.0]
 
 * Most of the details are here https://fhir-fli.github.io/fhir_fli_documentation/docs